feat: Modularize API into separate components
- Create modular API structure with Blueprint pattern
- Separate business logic into service classes:
  * SujetoService: Subject management operations
  * MedicionService: Measurement data processing
  * TaskLogService: Task logging operations
  * ExportService: CSV export functionality
- Organize routes by functionality with /api prefix
- Add centralized API configuration
- Reduce main app.py from 540+ to ~150 lines (72% reduction)
- Maintain backward compatibility with existing API contracts
- Add comprehensive API documentation

Closes: API modularization requirements
